// src/app/interface.ts
export interface smiplifiedproduct {
_id: string;
imageURL: string;
price: number;
slug: string;
categoryName: string;
name: string;
}
export interface fullproduct {
    _id: string;
    images: string;
    price: number;
    slug: string;
    categoryName: string;
    name: string;
    description: string;
    price_id: string; 
}